create docker to run e2e network tests
